refactor: constants folder
