+++ src/constants/banks/index.ts
@@ -1,0 +1,1 @@
+export { BANKS_DATA } from './banksData';
