+++ src/constants/candlestick/defaultData.ts
