+++ src/constants/contacts/index.ts
@@ -1,0 +1,2 @@
+export { CONTACTS_INFO } from './contactsInfo';
+export { SOCIAL_LIST } from './socialList';
