+++ src/constants/currencies/index.ts
@@ -1,0 +1,1 @@
+export { DEFAULT_CURRENCIES } from './defaultCurrencies';
