+++ src/constants/index.ts
@@ -1,13 +1,12 @@
 export { AppRoutes } from './appRoutes';
 export { AppTheme } from './appTheme';
-export { BANKS_DATA } from './banksData';
-export { CONTACTS_INFO } from './contactsInfo';
-export { DEFAULT_CHART_DATA } from './defaultChartData';
-export { DEFAULT_CURRENCIES } from './defaultCurrencies';
+export { BANKS_DATA } from './banks';
+export { CANDLESTICK_OPTIONS, DEFAULT_CHART_DATA } from './candlestick';
+export { CONTACTS_INFO, SOCIAL_LIST } from './contacts';
+export { DEFAULT_CURRENCIES } from './currencies';
 export { FOOTER_LINK_GROUPS } from './footerLinkGroups';
 export { ICONS } from './icons';
 export { PAGE_LIST } from './pageList';
-export { SOCIAL_LIST } from './socialList';
 
 export const CURRENCY_MODAL_QUERY_KEY = 'currency';
